requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
WTForms==3.1.1
python-dotenv==1.0.0
requests==2.31.0
reportlab==4.0.7
python-docx==1.1.0

